+++ lib/utils/execSync.js
@@ -1,0 +1,7 @@
+var child = require("child_process");
+
+module.exports = function execSync(cmd) {
+  return child.execSync(cmd, {
+    encoding: "utf8"
+  }).trim();
+};
