+++ lib/utils/getPackageConfig.js
@@ -1,0 +1,5 @@
+var getPackageLocation = require("./getPackageLocation");
+
+module.exports = function getPackageConfig(packagesLoc, name) {
+  return require(getPackageLocation(packagesLoc, name) + "/package.json");
+};
