+++ lib/utils/getPackageLocation.js
@@ -1,0 +1,3 @@
+module.exports = function getPackageLocation(packagesLoc, name) {
+  return packagesLoc + "/" + name;
+};
