+++ lib/utils/gitAddFile.js
@@ -1,0 +1,5 @@
+var execSync = require("./execSync");
+
+module.exports = function gitAddFile(file) {
+  execSync("git add " + file);
+};
