+++ lib/utils/progressBar.js
@@ -1,7 +1,7 @@
 var ProgressBar = require("progress");
 var pad         = require("pad");
 
-module.exports = function (total) {
+module.exports = function progressBar(total) {
   var bar = new ProgressBar(":packagename ╢:bar╟", {
     total: total,
     complete: "█",
@@ -12,7 +12,7 @@ module.exports = function (total) {
     width: (process.stdout.columns || 100) - 50 - 3
   });
 
-  return function (name) {
+  return function tick(name) {
     bar.tick({
       packagename: pad(name.slice(0, 50), 50)
     });
